# lib, vtysh: calculate the number of lines for better vtysh pagering.