It checks the lines of the output of the vty_out()s by buffering, and compare it with the terminal length before invoking the pager. Previously, the pager is always invoked regardless of the number of output lines when "terminal pager" is enabled. This is problematic if 1) "less" is invoked without "-F", or 2) "more" is used, in that the unnecessary paging lock occured even when there's no output. I imagine this is the reason why we can't enable terminal pager by default.

This patch replaces vty_out() (older is saved by the name: vty_out_old_pager()) and flushing the buffer is postponed to the vty_output_flush_vtysh_pager(), and when it is being flushed, the number of lines is calculated and used to decide whether the pager should be called. Also by this architecture change, we can remove FIXME about "exit" cmd, it enables better behaviors for "more" and "less" without "-F", and we might even make the "terminal pager" by default. Additionally, the pager is applied when help messages by "?" is longer than the screen.