@yasuhironet 2023-09-29
I fixed one important bug that corrupts the output of show commands in json.

After the fix, it comes down to just five BGP SRv6 L3VPN tests on Ubuntu 18 *ARM* have failed.
Those tests seem to be skipped in x86.

And now I am not sure about what should be done next...