@yasuhironet 2023-10-13
ci:rerun



2023年10月12日(木) 23:33 Donald Sharp ***@***.***>:

> @yasuhironet <https://github.com/yasuhironet> -> The Checkout code:
> Successful with additional warnings warnings need to be fixed from our
> perspective. Review will happen after that.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/FRRouting/frr/pull/14278#issuecomment-1759733964>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACDICZC6N7BN5NC5INGA42DX675TPANCNFSM6AAAAAA37FYS2I>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>