@yasuhironet 2023-10-14
hmmmm

failed: 
- Ubuntu 18.04 arm8 Part 4: bgp_lu_explicitnull.test_bgp_lu_explicitnull test_converge_bgplu
  E AssertionError: r1, prefix 192.168.2.2/32 from r2 not present
- Ubuntu 18.04 i386 Part 4: bgp_l3vpn_to_bgp_vrf.test_bgp_l3vpn_to_bgp_vrf test_check_scale_down
  E Failed: New core[s] found: /tmp/topotests/bgp_l3vpn_to_bgp_vrf.test_bgp_l3vpn_to_bgp_vrf/ce1/zebra_core-sig_6-pid_27856.dmp

(in search for a hint for debug)
Are these tests also tested (and succeeded) on other platforms in this CI run?

I changed my affiliation recently and can't access to any server for development anymore...