@donaldsharp 2023-11-01
```
BEFORE:

Displayed  944113 routes and 1888224 total paths

________________________________________________________
Executed in   54.06 secs   fish           external 
   usr time    3.69 secs    0.00 micros    3.69 secs 
   sys time   10.61 secs  1204.00 micros   10.61 secs 


—————————

AFTER:

Displayed  944102 routes and 1888202 total paths

________________________________________________________
Executed in   95.96 secs   fish           external 
   usr time    2.71 secs  1255.00 micros    2.71 secs 
   sys time    9.31 secs  419.00 micros    9.31 secs 

```

I decided to run this against 2 full bgp feeds that I have running locally.  Before these changes the timed run took 54 seconds after it takes 95 seconds.  

This is a NACK from me until performance is in the same ballpark